requirements.txt
matplotlib==3.6.1
pandas==1.5.0

pre-commit==2.20.0

pytest==7.1.3
pytest-cov==4.0.0
python-dotenv==0.21.0

requests==2.28.1

sqlalchemy==1.4.42
